Extract search input to separate component
